requirements.txt
fpdf2
streamlit
pypdf
Pillow
pytesseract